pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-image-mcp"
version = "1.0.0"
description = "MCP server for local image generation using Qwen-Image model"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "fastmcp>=2.0.0",
    "diffusers>=0.31.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "accelerate>=0.24.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "safetensors>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/qwen-image-mcp"
Issues = "https://github.com/yourusername/qwen-image-mcp/issues"

[project.scripts]
qwen-image-mcp = "server:main"

[tool.setuptools]
py-modules = ["server"]